Added Dark mode toggle button
